fix: allow collab revert even if ln channel is closed
By retrieving the channel signer through the channel monitor we are able to access the fund key of the original ln channel even though ldk closed and removed this channel.
Note: this patch includes a temporary dependency on rust-lightning. This will not be needed anymore once we upgrade to rust-lightning 116 because we can then access the channel keys directly.
Differently said, we might be able to drop this commit completely once we upgrade to rust-lightning 116

Signed-off-by: Philipp Hoenisch <philipp@coblox.tech>