feat: Add backup and restore
### Replace rust-dlc SledStorageProvider with DlcStorageProvider
This removes the dependency to the `dlc-sled-storage-provider` and the `simple-wallet`, however, it does not change the underlying implementation. The new DlcStorageProvider is simply introducing a new abstraction layer to allow for replacing sled with another implementation in preparation to hook in the backup.

### Add data dir to config
Note, I didn't put it on the environments on the client side as it requires the context to be initialized and that would have meant bigger changes, than I wanted to deal with.

### Verify users intent
The users intent is verified by verifying the signature on the backup and restore messages.

### Symmetric backup cipher
Backups are encrypted and decrypted on the client side.

### Add full backup support

If the app is upgraded to the new version it will run an initial full backup to ensure that the backup is complete.

- Combine states into a single state file
For testing I had to make the state mutable as otherwise I wasn't able to start a new app node.

### Improve the app onboarding flow

- The user should not be required to input the email when restoring from a backup. Note, the solution is a bit hacky, and ideally we also backup the preferences.
- Removed app bar from welcome screen
- Added visual improvements on the loading screen
- Wait for restore / initial backup to complete on loading screen
- Remove transitions from initial screens
- Remove splash screen once on the loading screen