fix: Load persisted DLC ChainMonitor whenever possible
If we cannnot load it because the deserialisation fails, we introduce
a workaround that ensures that we can overwrite it and move on with
the `DlcManager` initialisation.

This is generally not safe, but it unblocks us for the time being. We
should nevertheless remove the workaround as soon as most users have
updated to this patch.