feat: Allow position resizing via two-step protocol
With an open position, a trader is now able to place and execute
orders normally. Before they were limited to closing their open
position.

We achieve this by chaining a subchannel close protocol with a
subchannel open protocol. For this we have to keep some information
around before the two-step protocol is complete. This may be brittle
as we are eagerly updating the position just after commencing the
second step. We think this is acceptable because we already do the
same thing when opening a new position, but we should look to move
away from this pattern.

The app has had to change in a few ways:

- We now persist trades as opposed to periodically computing them for
the frontend. The implementation is therefore simpler.

- We've had to disable the leverage slider when placing orders with an
open position, as it doesn't make a lot of sense to change the
leverage through a new order (and it makes things way more
complicated).

- Trades are displayed as `X Long/Short` in the wallet history screen, as
opposed to `Opened/Closed position`. We have chosen to split trades
that change the direction of an open position into two separate
trades: one closing the original position and another one opening the
new one in the opposite direction. This is simpler to reason about,
but there might be potential to combine them if we really want to.