fix: Do not call periodic check twice
The periodic check is already called in a loop by the `sub_channel_manager_periodic_check`. I don't see a reason to do it multiple times.

Also finalizing a force-closure with the `period_check` called from the manager lead to a panic, when trying to get the attestations from the oracle.