fix: Reduce collaborative revert reliance on ChannelManager
In situations where the collaborative revert protocol might be useful,
it is likely that the channel will have been erased from the
`ChannelManager`.

Before this patch we had unavoidable dependencies on the
`ChannelManager`, but with this patch we are able to perform the
protocol without looking for the relevant `ChannelDetails` via the
`ChannelManager`.