fix(mobile): Compute trade cost correctly
Before, the trade cost (the main number associated with a trade in the
wallet history screen) did not include the PNL or the order-matching
fee. As such, it was only reflecting the margin change.

Now we actually do include these values in the calculation. These
values were already available as they were being displayed separately.