feat: Prevent resizing more than the current position quantity
While working on https://github.com/get10101/10101/issues/1625 I noticed that the way we do a resizing with changing the direction completely can't work the way we have it designed at the moment.

The issue is that we have a bid and an ask price. If we are changing the general direction of our position during resizing we are not using the correct price for the new position.

As a quick fix I've added a simple UI validation rule, that prevents the user from reducing a position further than the current quantity of the position.