fix(app): Keep same average_execution_price when reducing position
When the position was being partially reduced (e.g. from 10 long to 5
long), we were wrongly updating the average execution price of the
position. But removing contracts from a position should not affect the
contracts that remain in the position!

This bug resulted in showing incorrect values to the app's UI.