fix(coordinator): Calculate average execution price correctly
When resizing a position, the average execution price only changes if
the position is extended or if the position direction changes (the new
average execution price is equal to the execution price of the order
tha flipped the direction). If the position is reduced the average
execution price does not change.