fix(coordinator): Use PNL when closing channel for two-step resize
Before, when executing the two-step position resize protocol we
naively assumed that the DLC channel could be closed as if nothing
happened. This led to not honouring the PNL generated by the position
if the position was being reduced or if the position direction was
being flipped.

Now, we do include the PNL (if applicable) in the
`accept_settlement_amount` used to settle the first DLC channel before
remaking it for the purposes of resizing.