chore: Remove Default LnDlcNodeSettings
It is dangerous to assume that app, coordinator and maker all want the
same `LnDlcNodeSettings`. By removing the `Default` implementation it
is much clearer that each kind of node should choose different
settings.

For now we do not modify the values, so all of them still share the
same settings values.