chore(coordinator): Stop using default settings
We have wasted too much time in the past chasing bugs caused by the
coordinator defaulting the settings if the settings file was missing
or could not be parsed.

As devs, we lost the ability to start the coordinator locally using
`cargo run --bin coordinator` without making sure that there is a
valid settings file at the default data directory. But the `just
coordinator` command has been modified to handle this for us by
copying a `test-coordinator-settings.toml` into the default data
directory if needed.

We've also introduced a `prod-coordinator-settings.toml` to serve as
an example.