fix(ln-dlc-storage): Read and write ChainMonitor using the same key
This fixes a kind of bug which could happen like this:

1. Create a `ChainMonitor` based on version 1.4.4 (for example). This
would use the number "4" as the key.

2. Upgrade to version 1.6.5 (for example).

3. On startup, we would identify the original `ChainMonitor` as
backwards-incompatible and attempt to overwrite it. Instead, we would
insert it under the new key "chain_monitor".

4. Then we would build the `dlc_manager::Manager`, which would load
the `ChainMonitor` internally. This would look for the _first_
`ChainMonitor` out of all the ones in the `CHAIN_MONITOR` tree,
ignoring the key. Because the original one was inserted first, we
would load the original one which would once again fail to deserialise
due to backwards-incompatibility.

Now we ensure that we always use the same key to read and write the
`ChainMonitor`. This makes sense because there should only ever be one
`ChainMonitor` per node.