chore(tests-e2e): Reduce linking time by using Rust unit tests
Rust "unit" tests are just tests that are located next to the code. By
putting the e2e tests next to the auxiliary code in the `tests-e2e`
crate, we avoid re-linking and building a separate binary for each e2e
test.

Advice taken from this blog post[1], which we already applied to the
`ln-dlc-node` crate.

[1]: https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html.