chore: Resend authentication on hot restart
On hot restart we did not receive any initial data, leaving the app in an intermediate / broken state without the liquidity options.

This changes re-submits the authenticate message on hot restart, thus moving closer to the scenario as if the app has been restarted.

Note, this has only been added to make the development experience nicer.