fix: Reduce websocket reconnect timeout to 200ms
This was originally increased to ensure that a rollover is not failing if the app is open and the coordinator restarts while moving into the rollover window, as the rollover will fail.

However, this is a very edge case, that shouldn't happen very often and even if it's happening. The rollover will simply timeout and finish on the next restart of the app without any errors.

Reducing this timeout to 200ms gives the app a more ready feeling as the app will not wait 5s before reconnect and thus fetching the latest limit orders to get a price.