chore: only have open position when sub channel opened
So far we optimistically stored in the db the new position when we proposed the subchannel offer to the other party. This is prone to failures, because we might never get said accept.
Instead, we introduce a new state `Proposed`. A position is in proposed until the `Finalize` message has been received, i.e. until the subchannel (and the position) is actually open.