wip: reproduce timing issue
if a position is getting resized we close the current contract and reopen a new one. It might happen that the task  will pick it up in the meantime and set the  to closed. This breaks the renew protocl because we expect the position to be in a certain state