fix: race condition on resizing
By setting the position only to closed in our db when we receive the `CloseFinalize` event we can get rid of the async task `closed_positions` which checked regularly if a position needs to be closed in the db.
This helps us to remove a race condition where a position was accidentally set to closed which was actually in resizing.