fix: connection race condition
In `track_channel_status` we might need a connection to the coordinator. This is not always guaranteed because we spawn the connection task async as well. Hence, we wait here up to 1 minute. If we still haven't connected we give up and continue nevertheless. The below function call _might_ fail in certain cases but it's not the end of the world as we will retry simply when the user restarts the app the next time.
Hopefully by then the coordinator is reachable