fix(app): Ensure that we can only have one Filling Order at a time
If no new `Order`s can be submitted until the current `Order` in
`Filling` is either moved to `Filled` or `Failed`, we ensure that we
cannot have multiple `Order`s in `Filling` at the same time.

This is important because this was supposed to be an invariant which
we were not upholding.