fix(app): Handle possibility of multiple orders in Filling
The previous patch attempted to prevent the situation from happening,
but we have already encountered instances where this has happened.

To handle them, we choose to:

- Take the _oldest_ `Order` in `OrderState::Filling`, because it is
the one that was never resolved in the first place. This order is the
one most likely to match the unfinished update to the subchannel.

- Discard all other `Order`s in `OrderState::Filling`, because the
user should not have been able to create them anyway.