chore: refactor payout curve example
The example now produces a csv file which contains `<price>,<offerer>,<acceptor>`. This is useful to plot the payout for the coordinator(offerer) and trader (acceptor) in a single diagram.

I've also changed the gnuplot file which now produces an SVG so that one can zoom in and see more details.

Signed-off-by: Philipp Hoenisch <philipp@coblox.tech>