feat: Replace fee with collateral reserves in payout curve model
Originally, the `fee` determined the minimum amount that the
coordinator (the offering party) would get out of a DLC i.e. if they
got liquidated.

With the move away from LN-DLC channels, we still want to be able to
have part of the funds in the channel separate from the current bet.
Before, they used to sit in the Lightning channel. A stable `rust-dlc`
DLC channel can only have either a single DLC output (when
`Established`) or 1-2 party-owned output(s) (when `Settled`). This
means that there is no way to "split" the purpose of the coins in the
channel based on transaction outputs.

Because of this, we instead enhance our payout curve to be able to
effectively exclude part of the collateral in the DLC channel from the
current bet. This is the same mechanism that we used to use for the
aforementioned `fee`, but it has now been renamed to
`collateral_reserve` and it now can be applied to both parties (not
just the offering party).

It's actually better to use this virtual mechanism to "split" the
channel, because if we ever have to go on-chain with a `Established`
channel there is only one DLC output to spend, reducing transaction
fee costs significantly.