fix: Rework method to get DLC channel usable balance
This allows us to actually take into account how much money is part of
a DLC but not being traded.

The trick is to rely on the payout curve associated with the contract.
For each party, the lowest possible payout equals their collateral
reserve. Since those coins are not at risk, they can be used to trade
and so are usable.

I do think the overall logic should not be set in stone, particularly
around the non-stable channel and contract states, but this should be
a reliable start.