chore: fix reserved utxos
We need to remember utxos which were used already so that they are not being reused by accident.
We needed to do this because we don't want to sync with the blockchain so often.
So far we stored reserved utxos in a sled db and overwrote this by loading utxos from our bdk wallet.
Instead, we now store reserved utxos in memory only.
In memory means, it is gone on a restart.
This is not a problem, because on a restart we sync with the blockchain again and know if the utxo was spent or not.

Signed-off-by: Philipp Hoenisch <philipp@coblox.tech>