fix(coordinator): Allow requests with up to 50MB bodies
By default this is set to 2MB, according to the docs[1].

We need larger bodies than before (even in the e2e test!), because we
added lots of CET adaptor signatures to the database recently[2].

I also played around with nginx's `client_max_body_size`, but that's
actually irrelevant since we don't use nginx as a reverse proxy for
the coordinator in the e2e tests.

Related to https://github.com/get10101/10101/issues/1719.

[1]: https://docs.rs/axum/latest/axum/extract/struct.DefaultBodyLimit.html#method.disable.
[2]: https://github.com/get10101/10101/pull/1817.