fix(tests-e2e): Ensure that coordinator DLC channel is confirmed
Once we create the first DLC channel (with the first position), we
require a confirmation. Both the app and the coordinator need to see
the corresponding transaction get one confirmation and update their
DLC `Contract` states accordingly.

We were already doing this for the app by calling `sync_dlc_channels`.
And we are now also doing it for the coordinator by hitting the
`/api/admin/sync` endpoint.