fix(coordinator): Reject closing trade if channel not confirmed
We had introduced this check for the app, but not for the coordinator.
Obviously the app might know that the channel is confirmed, but the
coordinator doesn't have to agree.

In such a scenario, by failing early we get a much more understandable
error message and we prevent the renew protocol from moving along any
further than it needs to i.e. it doesn't actually start.