chore(coordinator): Do not duplicate error logs
When debugging a recent problem, I was looking for an ERROR message in
the coordinator logs. I saw these two ERROR messages and I was
immediately confused, until I finally realised that they were the same
error message, but duplicated.

```
ERROR Failed to execute trade. Error: Cannot trade with counterparty X with DLC channel in state "SettledConfirmed" trader_id=X order_id=Y
DEBUG back_up: coordinator::backup: Create user backup node_id=Z backup.key="10101/db"
ERROR error=InternalServerError("Could not handle trade request: Cannot trade with counterparty X with DLC channel in state \"SettledConfirmed\"")
```

That's not good developer experience. The reason why this was added
was to really make sure that error messages are printed to the logs,
but we should just fix the places where errors are being swallowed,
rather than doing something like this.

A couple interesting things I've done here:

- Use `with_context` on the error returned by `trade` so that we still
add the context of the `order_id` and `trader_id`, which was something
that was good about the original code.

- Remove trader ID and order ID context from a few spots in
`trade_internal`, since they are already part of the error returned by
its only caller i.e. `trade`.

- Match against the `update_order_and_match` error case and log
instead of bubbling up. This one is important because (1) we want to
see the error, but (2) we do not want to forget that the original
error is a different one! In general, when we are already on an error
path and we are doing some extra handling, we should not just bubble
up any new errors.