docs: Remove todo about first signed dlc channel
At the moment we can be certain that there will ever only be one **signed** dlc channel. A previous channel that has been closed on-chain is not in a signed state and will not be returned by this function.

Use `list_dlc_channels` to get also non-signed dlc channels. Note, its not safe to use `first` on the `list_dlc_channels` function.