fix(app): Map SignedChannelState to ChannelStatus correctly
I think it's correct to assume that a `Established` channel is one
that has a position. But a `RenewFinalized` channel is still pending
completion of the renew protocol and will eventually transition back
to `Established`.

This behaviour can be seen in our `can_open_and_settle_offchain` test
in the `ln-dlc-node` crate.