chore: Reject pending offers
If the app (on reconnect) identifies that there is a pending offer (dlc channel, settle offer or renew offer) it will automatically try to reject it

I think this is the safest we can do given that 1. for dlc channels the utxo might not be reserved anymore and 2. (more importantly) the price might have changed again, so we probably don't want to continue a dlc protocol that started some days ago.