chore: Prevent proposing another dlc channel offer if one is already pending
This is a safeguard as we don't want to offer multiple dlc channels, which could lead to weird behaviours. E.g. due to a recent bug, the user had multiple positions in `Proposed` and after finally succeeding all positions where set to open because we don't link the contract id yet (https://github.com/get10101/10101/issues/1921)