fix: Use counterparty available channel balance to validate trades
Before this patch, when we had an open channel we would take the
_original_ counterparty collateral as the funds that the counterparty
had available for trading. This is generally incorrect, as the
counterparty available funds vary with the state of the channel.

Since we already know how to calculate the available channel funds for
the party inquiring about this information, we can similarly calculate
it for their counterparty.

The calculations are very similar, and they have similar drawbacks:

- If we have more than one signed channel (should not happen, but
still), we take the aggregate funds of the lot. This is probably
desirable to display a total balance, but it doesn't work if we are
trying to execute trade, which is supposed to be executed in a single
channel (for the foreseeable future).
- If a channel is in the process of being closed but not yet
completely closed, the funds might still appear as available. Once
again, this is okay for a total balance where we don't want funds to
disappear. But these funds are obviously not available for trading
using the channel that is being closed.

These problems should be addressed incrementally with the model
redesign motivated by https://github.com/get10101/10101/issues/1921.

Co-authored-by: Richard Holzeis <richard@holzeis.me>