fix: Inserting channel opening parameters before posting new order to the coordinator
I am not exactly sure why this is fixing the issue, but my assumption is the following.

If we receive the order match (async via websockets) before the post_new_order http request has finished. We'd still hold an open connection and the write lock on the order.

I presume this leads to the order match processing to fail because of Could not load order from db: cannot acquire database connection: timed out waiting for connection. But I am not sure why that would result into the processing of the order match running into a timeout when trying to get a connection.

When the post_new_order finally finishes the lock is freed up. I guess afterwards the pool would return connections again but we are already stuck because we do not send the trade request anymore (since processing the match already failed)

Additionally I had to remove the order_id check on the submit_order_change_notifier, since it could happen that we don't have the id yet if the order match is already processed before the http post new order request finished. IMHO that should be fine because we only ever have one active order anyways and we don't have to separate them.

fixes #2055