chore(ln-dlc-node): Rewrite funding transaction test
Soon `rust-dlc` will enforce including at least one change output per
party in the funding transaction, so that either party can fee-bump
it.

The original `can_open_channel_with_min_inputs` test actually forced
`rust-dlc` to generate a change-output-less funding transaction. This
is valid for now, but it doesn't help to build on assumptions
that are going to change soon.

The new test simply checks that the funding transaction pays fees at the
expected fee rate. If we can trust `rust-dlc` to respect the fee rate
we give it, we can confidently display this information to a user who
is about to open a channel, so that they know the transaction fee cost
of their action.