fix(app): Rename reservedFeeSats to fundingTxFee
The value that we were displaying corresponded to half of the funding
transaction fee.

The fee reserve cannot be derived just by looking at the funding
transaction, as it is included in the funding transaction shared
output. Eventually, we should display the fee reserve in the
corresponding wallet history entry (and perhaps in a channel details
section too).