fix(app): Remove fee from total on confirmation bottom sheet
On the confirmation bottom sheet, we tell the user that some coins
will be locked up in the DLC channel. This message seems to imply that
these coins will still be owned by the user.

The margin and the collateral reserve make sense, but any fee paid or
reserved should probably be excluded. We could summarise this
information elswhere on the screen, but I don't think it is necessary
as all the information is already present separately.