fix: relaxed constraint on failed orders
So far, a failed order needed a reason when converting between types. However, we failed to do so when setting orders to failed. I've relaxed this constraint now. If there is no reason in the db, we assume an `Unknown` reason in the frontend.
At the same time, we ensure that we set a reason when setting all orders to failed.