chore: Ensure backwards compatibility for electrs CLI argument
Patch cd04992884b88d911c04b48c075e4c23883cfa0f changed the name of the
`esplora` CLI argument to `electrs` for correctness. But the change
was introduced in a backwards-incompatible manner.

This patch solves that issue by introducing an alias for the CLI
argument with the original name.

All this applies to both the coordinator and the webapp.