feat: position websocket
Over this websocket we are sending updates about new trades. This means, if a user opens a new position or closes a position, an update is sent out over this websocket. The update includes an incremenatl change but also the total position state as reference. A client listening on this can then reason on this data.