feat(app): Perform full sync if old wallet DB is present
With the recent upgrade to `bdk:1.0.0`, we dropped existing wallet DBs
in favour of a new format. This meant that a user would have to
manually populate the new wallet DB to be able to see their wallet
history and funds.

This is not a very good user experience, as at least some users would
be surprised by this. This is why we have decided to execute a full
wallet sync on startup if we find an old wallet DB. Once complete, we
delete the old wallet DB since the data is irrelevant and we do not
want to run a full sync on startup every time.