chore(channel offer): Reject a dlc channel offer if the processing fails.
Unlike rejecting a dlc channel offer if accepting fails, the `Channel::Offered` and `Contract::Offered` have not been stored to the dlc store.

Hence, there is no rust-dlc state to be cleaned up. Howerver, `rust-dlc` expects the `Channel::Offered` to be found by it's `dlc_channel_id`, which if it can't will fail the function.

This patch implements the reject soly on the application layer, but we might want to consider moving that code to rust-dlc`