fix(app): Check for old wallet data in correct location
I believe some old apps do have a `wallet` file that indicates that
the app should probably be migrated to the new wallet DB via a full
sync.

But this is not the case for fresh 1.9.0 apps. For those, we need to
consider the presence of the `on_chain` directory.