fix(app): Don't error when user is typing send amount
When the user is about to send on-chain and is typing the send amount,
we dynamically calculate a fee estimate so that we can update the
screen as they update the value.

This is nice, but the user will inevitably start by typing in small
amounts for which calculating a fee does not make economical sense.

We enhance the library code to handle this case explicitly by using a
error enum variant `EstimateFeeError::SendAmountBelowDust`. We can
then leverage this in the app backend to simply ignore this kind of
error.