feat(app): Run DLC channels checks during wallet refresh
In the past we were concerned about running all these tasks together,
but after the `bdk` upgrade and the pivot away from Lightning there
should be no problems. Also, any error that might occur during
`sync_dlc_channels` is simply logged and does not interfere with
updating the wallet history and balance.

A nice consequence of this change is that if a user triggers a wallet
refresh and a fund transaction or a collaborative close transaction is
found to be confirmed, the corresponding channel should update
automatically, ensuring that the app's state is up-to-date and ready
to keep trading.