chore: introduce fvm (flutter version manager)
This allows us to have different flutter versions for webapp and mobile.
It is needed, because the webapp can be built best with the current `beta` version. Using latest `stable` will break the webapp in firefox and we would need to fall back to a different renderer which comes at the cost of performance.

Signed-off-by: Philipp Hoenisch <philipp@coblox.tech>