chore(coordinator): Add order-matching fee to Position
The order-matching fee will accumulate all the order-matching fees for
a given position: opening order-matching fee; arbitrarily many resize
order-matching fees; and closing order-matching fee.

This information can be used to compute the collateral reserve when
resizing the position. The accumulated order-matching fee will be the
lower bound for the coordinator's collateral reserve.