fix: Add more resilience when getting the node from state
We can get a price before the node has started, since we are already connected to the coordinator.

If that happens the app would crash since the node state has not been set yet.

This change makes a couple of apis more resilient towards the node not being available yet and returns a logical default value.

IMO this is a safe and better approach than simply crashing the app.