fix(payout_curve): Ensure that liquidation payouts are sane
We were allowing the short liquidation payout to go under the short
party's collateral reserve, which was obviously wrong.

The payout curve is still a bit rough around the edges (literally).
This is because we have to make relatively big payout jumps to connect
the liquidation intervals with the adjacent intervals. I'm not sure
there is a good solution for this other than going back to the
ItchySats payout curve.