feat: Add coordinator liquidation price to positions
This is needed so we don't have to calculate the liquidation price when we check for liquidation.

Since we want to replace that with the margin call at 10% before liquidation, but still don't change existing liquidation prices, this will allow us to only apply the margin call to new positions.