feat: Allow resizing positions
- I fixed calculation bugs in our original implementation of position
resizing that we used for LN-DLC channels. This is why some assert
values changed.

- I reintroduced the resizing e2e tests, which are now much more
helpful because they assert on the specific values of position and
trade for the app. This can be further improved by asserting on the
coordinator side. The same pattern should eventually be applied to
most e2e tests.

- I had to add an `is_complete` column to the coordinator's `trade`
table to be able to record the `trader_realized_pnl_sat` ahead of
time.

Limitations:

- The app must have an order in `Filling` to be able to update the
position after we get a resizing channel renewal offer. This pattern
has caused problems in the past. We will probably fix this my sending
extra metadata with the `rust-dlc` offer message.