fix: Add reference id to dlc channel
The channel id is unfortunately not unique throughout the whole lifecycle of creating a channel.

The offered channel has a "temporary" channel id, which is different to the channel id. Using the reference id we can link those ids together and ensure that the `OfferedChannel` get removed once the correct channel id is known.