fix: Derive referral code from node id
This is not ideal, but since the referral code can be derived deterministically from the node id its safe to do and will ensure that the app will not fail to start.

This fixes an issue where the referral status is not yet set because the app hasn't processed the authenticated message from the coordinator yet.