chore: make use of self-hosted meme endpoint
hosting the memes on github works great but querying them from within the app is suboptimal because there are only a number of requests one is allowed to make.
Hence, it's better to host the memes ourselves.

I opted for hosting them on the mainnet vm because if this is down, the app doesn't work anyways.
Ideally though, we would use some form of CDN for this.