fix: Set quantity to zero when initiation order view with open position
If we don't have a position we default the quantity to the max quantity. If we have an open position we default the quantity to zero because the contracts will be defaulted to the open quantity.

The behaviour should be that in case of an open position the default contracts on the opposite direction is to close the position.