fix(app): Display correct custom fee data when sending bitcoin
Before this change, after choosing a custom fee rate we would show an
absolute fee of the same value, which was incorrect. This mistake
would be carried forward to the confirmation screen.

With this change, we now show the fee rate and the corresponding
absolute fee.

Additionally, we've simplified the the fee-picking flow. We can now
pick a pre-determined fee rate with a single click. The custom fee
flow is hidden behind a button which displays a dialog where a custom
value can be input. Making the custom fee flow less prominent makes
sense as it is more likely that a user will want to use one of the
pre-determined options.

We've also removed a bit of USDp dead code.