fix(app): Improve readiness of send-onchain-screen
1. The original requirement to define a send amount before estimating
fees wasn't helpful and it was preventing the `send-onchain-screen`
from looking ready when first opened.

2. The default initial send amount of 1000 sats wasn't realistic or
useful. Additionally, it wasn't being set properly because the `init`
async function was never called.

We now set the initial default amount to 0. A 0 amount is invalid, so
we disable the `Send` button by default if no amount is defined in a
scanned QR code. In such a situation it's impossible to anticipate the
intention of the user, so it's better to require user interaction
before enabling the `Send` button.