feat(coordinator): Apply a funding fee on rollover
The coordinator generates funding fee events as funding rates are
added to the database and applies them on rollover.

The app displays funding fee events in a dedicated `Trades` tab,
together with regular trades. The funding fee events are added as soon
as the app learns about them from the coordinator, which is usually
before they are resolved.

Additionally, the app will update the position after rolling over
based on the funding fee events that were resolved with the rollover.

Missing bits:

- Apply funding fee events on closing and resizing.

- Considering funding fee events when deciding if positions need to be
liquidated.

- Add some unit tests.

- Expanding e2e tests.

- Add fee rate to app's `Trades` tab.

- Display next fee rate in app's `Trade` screen.

Co-authored-by: Philipp Hoenisch <philipp@hoenisch.at>